+++ src/main/java/org/apache/lucene/backward_codecs/lucene99/Lucene99ScalarQuantizedVectorsWriter.java
@@ -60,7 +60,8 @@ import org.apache.lucene.util.VectorUtil;
 import org.apache.lucene.util.hnsw.CloseableRandomVectorScorerSupplier;
 import org.apache.lucene.util.hnsw.RandomVectorScorerSupplier;
 import org.apache.lucene.util.hnsw.UpdateableRandomVectorScorer;
-import org.apache.lucene.util.quantization.QuantizedByteVectorValues;
+import org.apache.lucene.util.quantization.BaseQuantizedByteVectorValues;
+import org.apache.lucene.util.quantization.LegacyQuantizedByteVectorValues;
 import org.apache.lucene.util.quantization.QuantizedVectorsReader;
 import org.apache.lucene.util.quantization.ScalarQuantizer;
 
@@ -214,8 +215,8 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
     }
 
     @Override
-    public void mergeOneField(FieldInfo fieldInfo, MergeState mergeState) throws IOException {
-        rawVectorDelegate.mergeOneField(fieldInfo, mergeState);
+    public void mergeOneFlatVectorField(FieldInfo fieldInfo, MergeState mergeState) throws IOException {
+        rawVectorDelegate.mergeOneFlatVectorField(fieldInfo, mergeState);
         // Since we know we will not be searching for additional indexing, we can just write the
         // the vectors directly to the new segment.
         // No need to use temporary file as we don't have to re-open for reading
@@ -242,20 +243,6 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
                 docsWithField
             );
         }
-    }
-
-    @Override
-    public CloseableRandomVectorScorerSupplier mergeOneFieldToIndex(FieldInfo fieldInfo, MergeState mergeState) throws IOException {
-        if (fieldInfo.getVectorEncoding().equals(VectorEncoding.FLOAT32)) {
-            // Simply merge the underlying delegate, which just copies the raw vector data to a new
-            // segment file
-            rawVectorDelegate.mergeOneField(fieldInfo, mergeState);
-            ScalarQuantizer mergedQuantizationState = mergeAndRecalculateQuantiles(mergeState, fieldInfo, confidenceInterval, bits);
-            return mergeOneFieldToIndex(segmentWriteState, fieldInfo, mergeState, mergedQuantizationState);
-        }
-        // We only merge the delegate, since the field type isn't float32, quantization wasn't
-        // supported, so bypass it.
-        return rawVectorDelegate.mergeOneFieldToIndex(fieldInfo, mergeState);
     }
 
     @Override
@@ -586,6 +573,15 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
         return null;
     }
 
+    private static LegacyQuantizedByteVectorValues getLegacyQuantizedVectorValues(QuantizedVectorsReader reader, String fieldName)
+        throws IOException {
+        final BaseQuantizedByteVectorValues values = reader.getQuantizedVectorValues(fieldName);
+        if (values instanceof LegacyQuantizedByteVectorValues legacyValues) {
+            return legacyValues;
+        }
+        return null;
+    }
+
     private static ScalarQuantizer getQuantizedState(KnnVectorsReader vectorsReader, String fieldName) {
         QuantizedVectorsReader reader = getQuantizedKnnVectorsReader(vectorsReader, fieldName);
         if (reader != null) {
@@ -696,7 +692,7 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
      */
     public static DocsWithFieldSet writeQuantizedVectorData(
         IndexOutput output,
-        QuantizedByteVectorValues quantizedByteVectorValues,
+        LegacyQuantizedByteVectorValues quantizedByteVectorValues,
         byte bits,
         boolean compress
     ) throws IOException {
@@ -867,10 +863,10 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
     }
 
     static class QuantizedByteVectorValueSub extends DocIDMerger.Sub {
-        private final QuantizedByteVectorValues values;
+        private final LegacyQuantizedByteVectorValues values;
         private final KnnVectorValues.DocIndexIterator iterator;
 
-        QuantizedByteVectorValueSub(MergeState.DocMap docMap, QuantizedByteVectorValues values) {
+        QuantizedByteVectorValueSub(MergeState.DocMap docMap, LegacyQuantizedByteVectorValues values) {
             super(docMap);
             this.values = values;
             iterator = values.iterator();
@@ -887,8 +883,8 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
         }
     }
 
-    /** Returns a merged view over all the segment's {@link QuantizedByteVectorValues}. */
-    static class MergedQuantizedVectorValues extends QuantizedByteVectorValues {
+    /** Returns a merged view over all the segment's {@link LegacyQuantizedByteVectorValues}. */
+    static class MergedQuantizedVectorValues extends LegacyQuantizedByteVectorValues {
         public static MergedQuantizedVectorValues mergeQuantizedByteVectorValues(
             FieldInfo fieldInfo,
             MergeState mergeState,
@@ -901,15 +897,19 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
                 if (MergedVectorValues.hasVectorValues(mergeState.fieldInfos[i], fieldInfo.name)) {
                     QuantizedVectorsReader reader = getQuantizedKnnVectorsReader(mergeState.knnVectorsReaders[i], fieldInfo.name);
                     assert scalarQuantizer != null;
+                    final ScalarQuantizer quantizationState = reader == null ? null : reader.getQuantizationState(fieldInfo.name);
+                    final LegacyQuantizedByteVectorValues legacyVectorValues = reader == null
+                        ? null
+                        : getLegacyQuantizedVectorValues(reader, fieldInfo.name);
                     final QuantizedByteVectorValueSub sub;
                     // Either our quantization parameters are way different than the merged ones
                     // Or we have never been quantized.
-                    if (reader == null || reader.getQuantizationState(fieldInfo.name) == null
+                    if (reader == null || quantizationState == null || legacyVectorValues == null
                     // For smaller `bits` values, we should always recalculate the quantiles
                     // TODO: this is very conservative, could we reuse information for even int4
                     // quantization?
                         || scalarQuantizer.getBits() <= 4
-                        || shouldRequantize(reader.getQuantizationState(fieldInfo.name), scalarQuantizer)) {
+                        || shouldRequantize(quantizationState, scalarQuantizer)) {
                         FloatVectorValues toQuantize = mergeState.knnVectorsReaders[i].getFloatVectorValues(fieldInfo.name);
                         if (fieldInfo.getVectorSimilarityFunction() == VectorSimilarityFunction.COSINE) {
                             toQuantize = new NormalizedFloatVectorValues(toQuantize);
@@ -922,10 +922,10 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
                         sub = new QuantizedByteVectorValueSub(
                             mergeState.docMaps[i],
                             new OffsetCorrectedQuantizedByteVectorValues(
-                                reader.getQuantizedVectorValues(fieldInfo.name),
+                                legacyVectorValues,
                                 fieldInfo.getVectorSimilarityFunction(),
                                 scalarQuantizer,
-                                reader.getQuantizationState(fieldInfo.name)
+                                quantizationState
                             )
                         );
                     }
@@ -1020,7 +1020,7 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
         }
     }
 
-    static class QuantizedFloatVectorValues extends QuantizedByteVectorValues {
+    static class QuantizedFloatVectorValues extends LegacyQuantizedByteVectorValues {
         private final FloatVectorValues values;
         private final ScalarQuantizer quantizer;
         private final byte[] quantizedVector;
@@ -1122,14 +1122,14 @@ public final class Lucene99ScalarQuantizedVectorsWriter extends FlatVectorsWrite
         }
     }
 
-    static final class OffsetCorrectedQuantizedByteVectorValues extends QuantizedByteVectorValues {
+    static final class OffsetCorrectedQuantizedByteVectorValues extends LegacyQuantizedByteVectorValues {
 
-        private final QuantizedByteVectorValues in;
+        private final LegacyQuantizedByteVectorValues in;
         private final VectorSimilarityFunction vectorSimilarityFunction;
         private final ScalarQuantizer scalarQuantizer, oldScalarQuantizer;
 
         OffsetCorrectedQuantizedByteVectorValues(
-            QuantizedByteVectorValues in,
+            LegacyQuantizedByteVectorValues in,
             VectorSimilarityFunction vectorSimilarityFunction,
             ScalarQuantizer scalarQuantizer,
             ScalarQuantizer oldScalarQuantizer