+++ src/main/java/org/opensearch/knn/index/codec/KNN1040Codec/Faiss1040ScalarQuantizedFlatVectorsReader.java
@@ -6,6 +6,7 @@
 package org.opensearch.knn.index.codec.KNN1040Codec;
 
 import org.apache.lucene.codecs.hnsw.FlatVectorsReader;
+import org.apache.lucene.codecs.hnsw.FlatVectorsScorer;
 import org.apache.lucene.index.ByteVectorValues;
 import org.apache.lucene.index.FloatVectorValues;
 import org.apache.lucene.util.hnsw.RandomVectorScorer;
@@ -44,8 +45,12 @@ public class Faiss1040ScalarQuantizedFlatVectorsReader extends FlatVectorsReader
      *                                              will be wrapped to implement {@code HasIndexSlice}
      */
     protected Faiss1040ScalarQuantizedFlatVectorsReader(final FlatVectorsReader lucene104ScalarQuantizedVectorsReader) {
-        super(lucene104ScalarQuantizedVectorsReader.getFlatVectorScorer());
         this.delegateFlatVectorsReader = lucene104ScalarQuantizedVectorsReader;
+    }
+
+    @Override
+    public FlatVectorsScorer getFlatVectorScorer(String field) throws IOException {
+        return delegateFlatVectorsReader.getFlatVectorScorer(field);
     }
 
     @Override